Refactor to Postprocessor
Other processors/extensions may generate headings, that might need to be
downgraded. A prime example is the built-in raw HTML processing, that
removes inline HTML processes the actual markdown, and then -- in a
preprocessor -- puts the removed raw HTML back.

To avoid this, the header downgrade logic has been moved to a
postprocessor with a low priority (5) to make sure it's the last thing
running.